+++ vignettes/overview.R
@@ -34,3 +34,11 @@ AIC(fit3)
 #compare with AIC of unimodal model
 AIC(fit_abundances(abundances))
 
+#fit a bimodal model to a species classification dataset
+#and calculate the number of the differet categories in each octave
+data(categ, package="gambin")
+fits2 = fit_abundances(categ$abundances, no_of_components = 2)
+#without species classification data
+d1 <- deconstruct_modes(fits2, dat = categ, peak_val = NULL, abundances = "abundances", 
+     species = "species", categ = "status", col.statu = c("green", "red", "blue"))
+